use joda time everywhere for consistency
